[skip ci] Refactore new command pattern
