+++ frontend/src/components/annotation/commands/update_instance_command.ts
@@ -100,6 +100,7 @@ export class UpdateInstanceCommand {
       );
       initializedInstance.remove_edges();
 
+      console.log('end inistialized', initializedInstance)
       return initializedInstance;
     }
     else{
