+++ frontend/src/components/vue_canvas/instances/Instance.ts
@@ -46,7 +46,8 @@ export class Instance {
   public pause_object: false
 
 
-  public get_instance_data(): object {
+  // Returns any just to avoid warnings in the new Command pattern related class. Need to be replaces with the interface
+  public get_instance_data(): any {
     /*
     * Specific instance types should add/remove fields to this object if required.
     * */
