+++ frontend/src/helpers/command/avalible_commands/index.ts
@@ -1,0 +1,9 @@
+import CreateInstanceCommand from "./create"
+import UpdateInstanceCommand from "./update"
+import DeleteInstanceCommand from "./delete"
+
+export {
+    CreateInstanceCommand,
+    UpdateInstanceCommand,
+    DeleteInstanceCommand
+};
