+++ frontend/src/helpers/command/command.ts
@@ -39,67 +39,3 @@ export abstract class Command {
     abstract execute(): void;
     abstract undo(): void;
 }
-
-export class CreateInstanceCommand extends Command {
-    execute() {
-        const id_list = this.get_instances_ids()
-        this.instance_list.get_all().forEach((existing_instance, index, array_of_instances) => {
-            const { id, creation_ref_id } = existing_instance.get_instance_data()
-            if (id_list.includes(id) || id_list.includes(creation_ref_id)) {
-                array_of_instances[index].soft_delete = false
-            }
-        });
-
-        this.instances.forEach((_, index, array_of_instances) => {
-            array_of_instances[index].soft_delete = false
-        })
-    }
-
-    undo() {
-        const id_list = this.get_instances_ids()
-        this.instance_list.get().forEach((existing_instance, index, array_of_instances) => {
-            const { id, creation_ref_id } = existing_instance.get_instance_data()
-            if (id_list.includes(id) || id_list.includes(creation_ref_id)) {
-                array_of_instances[index].soft_delete = true
-            }
-        });
-    };
-}
-
-export class UpdateInstanceCommand extends Command {
-    private initial_instances: Array<Instance>;
-
-    public set_initial_instances(initial_instances: Array<Instance>): void {
-        this.initial_instances = initial_instances.map(instance => this._copyInstance(instance));
-    }
-
-    execute() {
-        this.instances.map((instance, index) => {
-            this.instance_list.replace(instance, this.replacement_indexes[index])
-        })
-    }
-
-    undo() {
-        this.initial_instances.map((instance, index) => {
-            this.instance_list.replace(instance, this.replacement_indexes[index])
-        })
-    }
-}
-
-export class DeleteInstanceCommand extends Command {
-    execute() {
-        console.log("executing delete")
-        this.instances.map((instance, index) => {
-            instance.soft_delete = true;
-            this.instance_list.replace(instance, this.replacement_indexes[index])
-        })
-    }
-
-    undo() {
-        console.log("undo delete")
-        this.instances.map((instance, index) => {
-            instance.soft_delete = false;
-            this.instance_list.replace(instance, this.replacement_indexes[index])
-        })
-    }
-}
